Add github auth to /add /edit /delete

+++ src/teuthology_api/routes/presets.py
@@ -1,10 +1,12 @@
 from fastapi import APIRouter, HTTPException, Depends, Response
+from sqlalchemy.orm import Session
+import logging
+
+from teuthology_api.services.helpers import get_token
 from teuthology_api.models import get_db
 from teuthology_api.models.presets import PresetsDatabaseException
 from teuthology_api.models import presets as presets_model
 from teuthology_api.schemas.presets import PresetSchema
-from sqlalchemy.orm import Session
-import logging
 
 log = logging.getLogger(__name__)
 
@@ -31,7 +33,17 @@ def read_preset(username: str, db: Session = Depends(get_db)):
 
 
 @router.post("/add", status_code=200)
-def add_preset(preset: PresetSchema, db: Session = Depends(get_db)):
+def add_preset(
+    preset: PresetSchema,
+    db: Session = Depends(get_db),
+    access_token: str = Depends(get_token),
+):
+    if not access_token:
+        raise HTTPException(
+            status_code=401,
+            detail="You need to be logged in",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
     db_preset = presets_model.get_preset_by_username_name(
         db, username=preset.username, preset_name=preset.name
     )
@@ -39,13 +51,22 @@ def add_preset(preset: PresetSchema, db: Session = Depends(get_db)):
         raise HTTPException(
             status_code=400, detail=f"Preset of this username & name already exists."
         )
-    return presets_model.create_preset(db, preset)
+    return presets_model.create_preset(db, preset.model_dump())
 
 
 @router.put("/edit/{preset_id}", status_code=200)
 def update_preset(
-    preset_id: int, updated_data: PresetSchema, db: Session = Depends(get_db)
+    preset_id: int,
+    updated_data: PresetSchema,
+    db: Session = Depends(get_db),
+    access_token: str = Depends(get_token),
 ):
+    if not access_token:
+        raise HTTPException(
+            status_code=401,
+            detail="You need to be logged in",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
     try:
         return presets_model.update_preset(
             db, preset_id, updated_data.model_dump(exclude_unset=True)
@@ -58,7 +79,17 @@ def update_preset(
 
 
 @router.delete("/delete/{preset_id}", status_code=204)
-def delete_preset(preset_id: int, db: Session = Depends(get_db)):
+def delete_preset(
+    preset_id: int,
+    db: Session = Depends(get_db),
+    access_token: str = Depends(get_token),
+):
+    if not access_token:
+        raise HTTPException(
+            status_code=401,
+            detail="You need to be logged in",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
     try:
         presets_model.delete_preset(db, preset_id)
     except PresetsDatabaseException as exc:
